build: Only have one AssemblyInfo file
Previously we had AssemblyInfo.cs generated from AssemblyInfo.cs.in by
the configure script, and used when building with make, and a
AssemblyInfoStatic.cs with hardcoded version numbers, used when building
with the csproj.

As the compilation is now done with xbuild, the one referenced from the
csproj is always used, which can cause some confusion.

We now reference AssemblyInfo.cs in the csproj, still generated from
AssemblyInfo.cs.in, to make sure version numbers are always up-to-date.
But we add it to git, so that you can use the csproj directly from a git
clone, without need to first run configure.

The unique place to set the version numbers is now configure.ac, and it
will propagate to all the right places.