Remove CancellationToken
DBus doesn't support cancelling methods. A user may expect
cancelling a method stops it immediately, but in practice
he'll have to wait for the method to complete.

Cancelling a method would only do something when there
were many parallel requests which isn't a common use-case.

We throw NotSupportedException when user tries to add CancellationToken to DBus method or signal.