Add MatchHeader for matching signals to callbacks
Since adding additional match types to SignalMatchRule, we can no longer
rely on an exact match (testing objects for equality). The HandleSignal
method in the DBusConnection class only has access to the message header
since the body has not been parsed yet, so we can only match on the
header fields at this point. Signal watchers that use any of the arg
matches will have to perform additional filtering later once the body
is decoded.