VariantValue: support constructing variant values. (#316)
The VariantValue type was created as a representation of a
variant that is read from the bus.

For writing variants, a separate type ('Variant') exists.

This PR makes it possible to compose variant values using
VariantValue. These values can then be written to the bus.

The UX of VariantValue when dealing with composite values
(Dictionaries, Arrays of Arrays/Dictionaries/Structs)
is clunkier than Variant (which offers a strongly-typed API).
Such composite values are not expected to be the common case.

The benefit of using VariantValue over Variant is that
values read from the bus as VariantValue can now be also
written back.